+++ packages/kysely/src/sessions/create.ts
@@ -9,8 +9,6 @@ export function create(db: Kysely<Database>) {
   ): Promise<Session> => {
     const createdSession = {
       ...session,
-      // fallback untill we changed primary key
-      session_id: session.id,
       created_at: new Date().toISOString(),
       updated_at: new Date().toISOString(),
       authenticated_at: new Date().toISOString(),